requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
boto3==1.39.7
botocore==1.39.7
Brotli==1.1.0
cached_path==1.7.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cn2an==0.5.22
colorama==0.4.6
contourpy==1.3.2
coqpit-config==0.2.1
coqui-tts==0.27.0
coqui-tts-trainer==0.3.1
csvw==3.5.1
cycler==0.12.1
Cython==3.1.2
datasets==4.0.0
dateparser==1.1.8
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.8
Distance==0.1.3
dlinfo==2.0.0
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
eng_to_ipa==0.0.2
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.3.0
fugashi==1.3.0
g2p-en==2.1.0
g2pkk==0.1.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.38.0
gradio_client==1.11.0
groovy==0.1.2
grpcio==1.73.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
inflect==7.0.0
isodate==0.7.2
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonlines==1.2.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langid==1.1.6
language-tags==1.2.0
lazy_loader==0.4
librosa==0.9.1
llvmlite==0.44.0
loguru==0.7.2
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mecab-python3==1.0.9
-e git+https://github.com/myshell-ai/MeloTTS.git@209145371cff8fc3bd60d7be902ea69cbdb7965a#egg=melotts
monotonic-alignment-search==0.2.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
networkx==2.8.8
nltk==3.9.1
num2words==0.5.12
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
orjson==3.11.0
packaging==25.0
pandas==2.3.1
phonemizer==3.3.0
pillow==11.3.0
plac==1.4.5
platformdirs==4.3.8
pooch==1.8.2
proces==0.1.7
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pykakasi==2.2.1
pyparsing==3.2.3
pypinyin==0.50.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
resampy==0.4.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.3
s3transfer==0.13.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
segments==2.3.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.47.1
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.13.3
tomlkit==0.13.3
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.27.4
triton==3.3.1
txtsplit==1.0.0
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==0.10.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
